requirements.txt
pyserial
elasticsearch
